Introduce a new functionality for the MapView.
